+++ src/main/java/org/sonatype/goodies/packageurl/MissingComponentException.java
@@ -20,7 +20,14 @@ package org.sonatype.goodies.packageurl;
 public class MissingComponentException
     extends InvalidException
 {
+  private final String name;
+
   MissingComponentException(final String name) {
     super("Missing required component: " + name);
+    this.name = name;
+  }
+
+  public String getName() {
+    return name;
   }
 }
